requirements.txt
wheel
djangorestframework
django-filter  для фильтрации
# swagger
drf-yasg  для свагеров(документации)
djangorestframework-simplejwt  для токенов
python-decouple
Pillow
psycopg2-binary
python-slugify  для слагов
celery  для быстрой отправки всего 
redis